% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSaver_dependency.R, R/fabric_dependency.R,
%   R/jQuery_dependency.R
\name{filesaver_dependency}
\alias{filesaver_dependency}
\alias{fabric_dependency}
\alias{jquery_dependency}
\title{Create an HTML dependency for FileSaver.js}
\usage{
filesaver_dependency()

fabric_dependency()

jquery_dependency()
}
\description{
Create an HTML dependency for FileSaver.js

Create an HTML dependency for fabric.js

Create an HTML dependency for jQuery
}
