% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fabric.R
\name{use_fabric}
\alias{use_fabric}
\title{Activate the fabricjs library}
\usage{
use_fabric()
}
\value{
called for the side effect of activating the fabricjs library
}
\description{
This function activates the javascript library fabricjs. You should put it at the beginning of your Rmd document or Shiny UI.
Note that you'll need to be connected to the internet.
}
